refactor - extract class
We move responsibility from fetching the rates to a separate class outside of the rate provider
