+++ jobs/Backend/Task/ExchangeRateUpdater/ICzechNationalBankExchangeRateGateway.cs
@@ -1,0 +1,6 @@
+﻿namespace ExchangeRateUpdater;
+
+public interface ICzechNationalBankExchangeRateGateway
+{
+    
+}
